requirements.txt
aiogram
asyncio
peewee
requests
celery
redis
tqdm
postgres
python-dateutil
datetime
paramiko
flask